Clean up branch
This commit simply changes permissions of files, moves license and
README file to top-level directory, and deletes extra files in the
java jail used for debugging purposes.  There should be no functional
changes as a result of this commit, so it wasn't tested.  Go back here
if bad things happen unexpectedly.