Initial Python support
This commit adds support for Python 2.7.  I have created a chroot jail
and successfully run a Python program that outputs the correct values.
The only interesting change in the entire source lies in cronScript.php
where I had to add a HOME=/home/contest value to the environment
variables for the chroot environment.  I'm not sure why this was
necessary since I was able to run the chroot command by hand without an
issue (even setting HOME=""), but when run through the cron job system,
things broke and this was the easiest fix.  The problem had to do with
looking in the right directory for user installed Python modules, so
this was the quick fix that definitely shouldn't be breaking anything
else.