Pass the great update of 2016
I updated the server after the 2016 CCSC:MW contest, and finally got it
working now.  These are the changes necessary to get the system running
because the upgrade included a newer, more strict, version of PHP.
Hopefully that additionally strictness identified a couple bugs that
have been fixed in the process.  I found a couple last night in
cronScript.php that looked a bit suspicious.  Look forward to additional
updates to incorporate newer versions of gcc/g++, Java, and potentially
a Python setup.