Update to newest chroot wrapper
This has been under development for quite some time and in very simple
tests appears to work.  Now is the time to flush out any problems with
the script as this will make better log files so that programs don't
overwrite old logs because they have the same name.  The source is also
much, much easier to read.