Add Exceeded Max Output feature
This is a new feature that is enabled by the new chroot wrapper.  Since
we use rlimit, we can tell by the signal when a file length is exceeded.
We now present this as a new response value and the following commit
supports it.  There is one hack in that for whatever reason, the signal
that is trying to be passed from the chroot wrapper to the PHP driver
does not succeed.  The hack hard codes the signal value because of this!
Ick.