Update hard-coded 9 for ACCEPT
Most of these changes are to update the hard-coded 9 that represents
ACCEPT to become a hard-coded 10.  The only other changes are a typo fix
and a query that didn't sort student submissions properly when building
the standings.  The correct final total showed, but sometimes not all of
the submission counts displayed.  Fixed it while finding this other bug.